feat:admin can see all orders from each user, and their details; controller is accessible from another classes;refactor: controller that was set in scene builder removed
This is needed in admin scene. There is table with all users which has "details" button in each row. On click there are all orders from one user and this is ALMOST the same situation as user watching his own orders. That's why controller must be initialized with some parameters and that can't be done when default one is set.