feat: admin can see and change user password
Like a regular user, an administrator must enter a valid, strong password when changing a user's password. To do this, open the password display button, then confirm that the user is an administrator (they must re-enter their password), and then click the change user password button. This will bring him to the same password change alert as the user